@@ -7,7 +7,7 @@
     "serve": "vite preview"
   },
   "dependencies": {
-    "vue": "^2.7.0",
+    "vue": "^3.5.13",
     "axios": "^0.20.0",
     "bootstrap": "4.6.0",
     "jquery": "^3.7.1"
@@ -15,6 +15,6 @@
   "devDependencies": {
     "typescript": "^5.0.0",
     "vite": "^4.0.0",
-    "@vitejs/plugin-vue2": "^2.0.0"
+    "@vitejs/plugin-vue": "^4.0.0"
   }
 }
