@@ -1,4 +1,4 @@
-import Vue from 'vue'
+import Vue, {createApp} from 'vue'
 import App from './App.vue'
 
 import 'bootstrap/dist/css/bootstrap.min.css'
@@ -6,8 +6,4 @@ import 'bootstrap/dist/js/bootstrap.min.js'
 import 'jquery/dist/jquery.min.js'
 import '@/assets/styles/global.css'
 
-Vue.config.productionTip = false
-
-new Vue({
-    render: h => h(App),
-}).$mount('#app')
+createApp(App).mount('#app')
